### 2020-06-11 #2323 @whedon
[PRE REVIEW]: pyomeca: An Open-Source Framework for Biomechanical Analysis
**Submitting author:** @romainmartinez (<a href="http://orcid.org/0000-0001-9681-9448">Romain Martinez</a>)
**Repository:** <a href="https://github.com/pyomeca/pyomeca/" target ="_blank">https://github.com/pyomeca/pyomeca/</a>
**Version:** 2020.0.1
