### 2020-06-15 2d11941
Added __version__
