+++ pyomeca/_version.py
@@ -1,0 +1,1 @@
+__version__ = "2020.0.2"
