### 2020-06-19 #2323 @Kevin-Mattheus-Moerman
@modenaxe @moorepants @ekjchadwick @Juanfrandepatech would you be interested in reviewing this submission for JOSS?

The review is on GitHub and focuses on [the software](https://github.com/pyomeca/pyomeca/) and this [short paper](https://github.com/openjournals/joss-papers/blob/joss.02323/joss.02323/10.21105.joss.02323.pdf).
