+++ .github/workflows/conda-package.yml
@@ -31,7 +31,7 @@ jobs:
           conda-channels: conda-forge
 
       - name: Add path DLL to windows path (windows only)
-        if: startsWith(matrix.os, 'windows')
+        runs-on: windows-latest
         run: echo "##[add-path]C:\Miniconda\Library\bin"
 
       - name: Install Dependencies
@@ -40,9 +40,13 @@ jobs:
           pip install -r requirements-dev.txt
           conda info -a
 
-      - name: Lint (ubuntu only)
-        if: startsWith(matrix.os, 'ubuntu')
+      - name: Lint python (ubuntu only)
+        runs-on: ubuntu-latest
         run: black --check .
+
+      - name: Lint (ubuntu only)
+        runs-on: ubuntu-latest
+        uses: docker://github/super-linter:v2.2.0
 
       - name: Test
         run: |
@@ -50,7 +54,7 @@ jobs:
           python -m pytest --cov-report term-missing --color=yes --cov=pyomeca tests
 
       - name: Coverage (ubuntu only)
-        if: startsWith(matrix.os, 'ubuntu')
+        runs-on: ubuntu-latest
         env:
           COVERALLS_REPO_TOKEN: XiFSzZ4JyITqinzFkJhNKD201jE7Bilmh
         run: |
