+++ docs/about.md
@@ -38,17 +38,17 @@ journal = {Journal of Open Research Software}
 }
 ```
 
-## Papers citing Pyomeca
+## Papers using Pyomeca
 
 - Blache, Yoann, Romain Martinez, Raphaël Dumas, Mickael Begon, Nicola Hagemeister, and Sonia Duprey. 2019. “[Chapter 20 - Motion Analysis and Modeling of the Shoulder: Challenges and Potential Applications.](https://www.sciencedirect.com/science/article/pii/B9780128167137000209)” In DHM and Posturography, edited by Sofia Scataglini and Gunther Paul, 261–271. Academic Press. doi:10.1016/B978-0-12-816713-7.00020-9.
 
-- Bouffard, Jason, Romain Martinez, André Plamondon, Julie N. Côté, and Mickaël Begon. 2019. “[Sex Differences in Glenohumeral Muscle Activation and Coactivation during a Box Lifting Task.](https://www.tandfonline.com/doi/abs/10.1080/00140139.2019.1640396)” Ergonomics 62 (10). Taylor & Francis: 1327–1338. doi:10.1080/00140139.2019.1640396.
+- Bouffard, Jason, Romain Martinez, André Plamondon, Julie N. Côté, and Mickaël Begon. 2019. “[Sex Differences in Glenohumeral Muscle Activation and Coactivation during a Box Lifting Task.](https://doi.org/10.1080/00140139.2019.1640396)” Ergonomics 62 (10). Taylor & Francis: 1327–1338. doi:10.1080/00140139.2019.1640396.
 
-- Martinez, Romain, Najoua Assila, Etienne Goubault, and Mickaël Begon. 2020. “[Sex Differences in Upper Limb Musculoskeletal Biomechanics during a Lifting Task.](https://www.sciencedirect.com/science/article/abs/pii/S0003687020300673)” Applied Ergonomics 86 (July): 103106. doi:10.1016/j.apergo.2020.103106.
+- Martinez, Romain, Najoua Assila, Etienne Goubault, and Mickaël Begon. 2020. “[Sex Differences in Upper Limb Musculoskeletal Biomechanics during a Lifting Task.](https://doi.org/10.1016/j.apergo.2020.103106)” Applied Ergonomics 86 (July): 103106. doi:10.1016/j.apergo.2020.103106.
 
-- Martinez, Romain, Jason Bouffard, Benjamin Michaud, André Plamondon, Julie N. Côté, and Mickaël Begon. 2019. “[Sex Differences in Upper Limb 3D Joint Contributions during a Lifting Task.](https://www.tandfonline.com/doi/abs/10.1080/00140139.2019.1571245)” Ergonomics 62 (5): 682–693. doi:10.1080/00140139.2019.1571245.
+- Martinez, Romain, Jason Bouffard, Benjamin Michaud, André Plamondon, Julie N. Côté, and Mickaël Begon. 2019. “[Sex Differences in Upper Limb 3D Joint Contributions during a Lifting Task.](https://doi.org/10.1080/00140139.2019.1571245)” Ergonomics 62 (5): 682–693. doi:10.1080/00140139.2019.1571245.
 
-- Assila Najoua, Claudio Pizzolato, Romain Martinez, David Lloyd, Mickael Begon. 2020. “Toward a physiological model of the shoulder: EMG-assisted algorithm to account for the muscle co-contraction.” Annals of Biomedical Engineering.
+- Assila Najoua, Claudio Pizzolato, Romain Martinez, David Lloyd, Mickael Begon. 2020. “[EMG-Assisted Algorithm to Account for Shoulder Muscles Co-Contraction in Overhead Manual Handling.](https://doi.org/10.3390/app10103522)” Applied Sciences.
 
 ## Contributors and support
 
