### 2020-07-03 #2323 @Kevin-Mattheus-Moerman
@mitkof6 we tend to aim for 2 weeks but we can be flexible
