### 2020-07-03 #2431 @whedon
[REVIEW]: pyomeca: An Open-Source Framework for Biomechanical Analysis
**Submitting author:** @romainmartinez (<a href="http://orcid.org/0000-0001-9681-9448">Romain Martinez</a>)
**Repository:** <a href="https://github.com/pyomeca/pyomeca/" target ="_blank">https://github.com/pyomeca/pyomeca/</a>
**Version:** 2020.0.1

@BKillen05 & @ @mitkof6, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
