### 2020-07-05 #2431 @Kevin-Mattheus-Moerman
@romainmartinez can you check the reference mitkof6 mentioned in this error message? You can run `@whedon generate pdf` here to check if the paper compiles.
