### 2020-07-06 #2431 @romainmartinez
> @romainmartinez can you check the reference mitkof6 mentioned in this error message? You can run @whedon generate pdf here to check if the paper compiles.

@Kevin-Mattheus-Moerman I don't have any references for `mitkof6`, which seems to be the reviewer's username.

In the current issue (#2431), @mitkof6 is displayed with two `@` while @BKillen05 has only one.
Maybe the problem is here?
