### 2020-07-07 #2431 @Kevin-Mattheus-Moerman
@openjournals/dev :point_up: I think this might be an interesting case to analyze. Somehow the system assigned the reviewer `@mitkof6` as `@ @mitkof6`, despite everything looking fine over at https://github.com/openjournals/joss-reviews/issues/2323#issuecomment-653506616. This surfaced when we got a paper compilation error and `whedon` oddly picked it up as a reference error in paper compilation. I was able to unassign the double @ version and assign the correct reviewer but this does seem like a funny kink in the cable somewhere.
