### 2020-07-07 #2431 @Kevin-Mattheus-Moerman
@BKillen05 @mitkof6 @romainmartinez apologies for a minor bug in our system. It is now resolved and the draft paper is available here:  :point_right: [Check article proof :page_facing_up:](https://github.com/openjournals/joss-papers/blob/joss.02431/joss.02431/10.21105.joss.02431.pdf) :point_left:
