### 2020-07-11 pyomeca/pyomeca#103 @mitkof6
Proper citation in JOSS
Dear authors,

I think the following publications are not properly cited:

- Hachaj, T., & Ogiela, M. R. (2019). RMoCap: An R language package for processing and
  kinematic analyzing motion capture data. Multimedia Systems. -> https://link.springer.com/article/10.1007/s00530-019-00633-9

- Hoyer, S., & Hamman, J. J. (2017). Xarray: N-D labeled arrays and datasets in python.
  Journal of Open Research Software, 5, 304. https://openresearchsoftware.metajnl.com/article/10.5334/jors.148/

- Muller, A., Pontonnier, C., Puchaud, P., & Dumont, G. (2019). CusToM: A matlab toolbox
  for musculoskeletal simulation -> https://hal.inria.fr/hal-01988715/document

I can follow the links provided and copy the "how to cite". Also, I find it useful to include the DOI if it exists. Please check and follow the citation rules of JOSS.

Thanks!

Dimitar
