### 2020-07-11 #2431 @mitkof6
https://github.com/pyomeca/pyomeca/issues/103
