### 2020-07-12 bdd46d1
Fix bibtex citations (DOIs) #103
