### 2020-07-13 pyomeca/pyomeca#103 @romainmartinez
Hello @mitkof6 and thank you for pointing out these mistakes.

PR #104 should have addressed this issue, let me know if it's good for you.
