### 2020-07-13 #2431 @romainmartinez
> pyomeca/pyomeca#103

Should be fixed in [#104](https://github.com/pyomeca/pyomeca/pull/104), thank you.
