### 2020-07-18 pyomeca/pyomeca#105 @mitkof6
Review pyomeca: future directions
Dear authors,

Thanks a lot for the excelent documentation and setting up everything in accordance with the software best practices.

I want to better understand your future directions and vision about this project.

The API and the use cases seem nice, however, in terms of added value it does not bring much to the user except basic operations such as reading, writing, filtering and plotting. **I would find it very useful, if this framework goes a step forward and provide functions for processing and extracting meaningful clinical outputs from the raw data.** For example, it would be very nice if you provide common utilities for extracting the kinematics and kinetics ensembles per gait cycle, or calculate cadence, speed, etc.

On the other hand, by looking at the repositories under pyomeca it seems that you are working on some nice packages that try to manipulate c3d files or interface with OpenSim. **Is pyomeca only related to the features that were described in the publication or is it going to be something bigger where you will provide additional modules in the future? Please indicate your vision about this project and things you plan to support in the future. Are the projects under [pyomeca](https://github.com/pyomeca) intended for this publication or they will be published in the future?**
