### 2020-07-18 #2431 @mitkof6
https://github.com/pyomeca/ezc3d/issues/108
