### 2020-07-22 pyomeca/pyomeca#106 @BKillen05
Review: Running EMG Example
Using the instructions on the repository I could not view the figures without first importing pyplot from matplotlib. Not sure if this was only a problem for me. System/Anaconda spec : Python 3.6.10, Windows 10 Enterprise, 64-bit
