### 2020-07-22 pyomeca/pyomeca#107 @BKillen05
Review: Additional examples for other data
Dear authors,

I appreciated the documentation and associated paper. The simplicity of the implementation and seemingly overlapping functions and methods for the different data types is useful for the beginning.

I have some small comments relating to documentation which could help the lower level user adapt and utilise the project.

First, it would be helpful to also be able to replicate the examples in your research paper as they are slightly different to the example you provide on the repository.

While the API functions were very well documented,  there are two interrelated cases I think would greatly benefit the project.
1. Not being familiar with the xrarry python type I had to search a little while to be able access individual data, i.e., the specific EMG channel , this would be handy for those wanting to look at a specific channel for example

2. The examples provided on the repository as well as within your research paper are centered around EMG data. Given you present the toolbox as a complete package for all biomechanical data - some additional example for Markers and Angles would be nice to see.
