### 2020-08-05 #2431 @Kevin-Mattheus-Moerman
@romainmartinez can you give an update on progress in relation to responding to the issues raised?
