### 2020-08-05 #2431 @romainmartinez
Sorry for the delay @Kevin-Mattheus-Moerman , I plan to address the issues raised before the end of the week.

I'll update the status of the issues here:

- :heavy_check_mark: [Done in PR [#136](https://github.com/pyomeca/ezc3d/pull/136)] [Make install requires sudo (Linux) for copying Python bindings](https://github.com/pyomeca/ezc3d/issues/108) by @mitkof6
- :white_check_mark: [Waiting for [PR #108](https://github.com/pyomeca/pyomeca/pull/108) to be merged] [Proper citation in JOSS](https://github.com/pyomeca/pyomeca/issues/103) by @mitkof6
- :construction: [Work in progress] [Future directions](https://github.com/pyomeca/pyomeca/issues/105) by @mitkof6
- :construction: [Work in progress] [Running EMG Example](https://github.com/pyomeca/pyomeca/issues/106) by @BKillen05
- :construction: [Work in progress] [Additional examples for other data](https://github.com/pyomeca/pyomeca/issues/107) by @BKillen05
