### 2020-08-05 pyomeca/pyomeca#103 @romainmartinez
Hello @mitkof6, is this OK?
