### 2020-08-05 pyomeca/pyomeca#106 @romainmartinez
Hello @BKillen05 and thanks for the issue.

`pyomeca`'s documentation uses [Jupyter notebooks](https://jupyter.org/) to show code and instructions.
If you use the server we provide [here](https://mybinder.org/v2/gh/pyomeca/pyomeca/master?filepath=notebooks), the figures appear correctly.
This process is quite standard and other libraries use it in their documentation: [`scikit-learn`](https://scikit-learn.org/stable/visualizations.html) (42k stars on GitHub), [`pandas`](https://pandas.pydata.org/docs/user_guide/10min.html#plotting) (26k stars on GitHub) or [`xarray`](http://xarray.pydata.org/en/stable/plotting.html) (1.7k stars on GitHub).
Under the hood, pyomeca uses `matplotlib`.
To show a figure in a non notebook environment, the user just has to add `plt.show()`.
