### 2020-08-07 pyomeca/pyomeca#103 @mitkof6
@romainmartinez it is ok!
