### 2020-08-10 pyomeca/pyomeca#106 @BKillen05
Hello @romainmartinez
Thank you for your reply.  I was running the example on a standalone anaconda environment not the Jupyter notebook you indicated. This is likely the problem as you say. It's possible I missed this but it is worth double checking this is documented
