### 2020-08-19 pyomeca/pyomeca#105 @mitkof6
Hi @romainmartinez,

Thanks a lot for your detailed response. I agree with you and I like your arguments. Please iterate over the manuscript one more time and include some of the points in your response. In particular, please improve the statement of need mentioning your response to the question:

> The API and the use cases seem nice, however, in terms of added value it does not bring much to the user except basic operations such as reading, writing, filtering and plotting.

In the summary, you can include your philosophy on having simple and powerful modules that are modular and easy to develop.

If you agree, it would be nice to have a conclusion section where you will outline your vision for this project and how it might evolve. It would be nice to mention how pyomeca will interact with other modules such as pyosim or biorbd-viz.

Thanks!
