### 2020-08-20 pyomeca/pyomeca#107 @romainmartinez
> I appreciated the documentation and associated paper. The simplicity of the implementation and seemingly overlapping functions and methods for the different data types is useful for the beginning.

We would like to thank you for those kind words.

> First, it would be helpful to also be able to replicate the examples in your research paper as they are slightly different to the example you provide on the repository.

We have kept exactly the same example as the one presented in the documentation [here](https://pyomeca.github.io/getting-started/#application-electromyographic-pipeline).
The only difference is in the first few lines when we do the muscle selection.
We wanted to omit these lines because they are entirely dependent on the file you are working with and we wanted to keep the example short for the paper.
It would have been difficult to distribute a `c3d` file with the paper so we chose to work with a dummy file that already contains the desired muscles.
Let us know if you want to add these lines to the paper example so that the two examples are exactly the same.

> Not being familiar with the xarray python type I had to search a little while to be able access individual data, i.e., the specific EMG channel , this would be handy for those wanting to look at a specific channel for example

We agree that the data structure requires some instructions.
The documentation contains instructions for [accessing the data](https://pyomeca.github.io/getting-started/#indexing), [accessing metadata](https://pyomeca.github.io/getting-started/#metadata), [doing calculations](https://pyomeca.github.io/getting-started/#computation) and graphics, etc.
We repeatedly mention the use of  `xarray` and refer users to their documentation, which contains much more information for manipulating the data.

> The examples provided on the repository as well as within your research paper are centered around EMG data. Given you present the toolbox as a complete package for all biomechanical data - some additional example for Markers and Angles would be nice to see.

We completely agree and we will add the following example in the documentation:

This example consists of calculating the elbow joint angle from skin markers. It uses `Markers`, `Angles` and `Rototrans`.

```python
rt_arm_in_global = Rototrans.from_markers(...)
rt_lower_arm_in_global = Rototrans.from_markers(...)

rt_arm_in_global_T = Rototrans.from_transposed_rototrans(rt_arm_in_global)
rt_lower_arm_in_arm = rt_arm_in_global_T * rt_lower_arm_in_global
angles = Angles.from_rototrans(rt_lower_arm_in_arm, seq)
angles.plot()
```
