### 2020-08-20 #2431 @romainmartinez
> @romainmartinez any more updates point_up ?

I still have two more points to address among the [issues](https://github.com/openjournals/joss-reviews/issues/2431#issuecomment-669236807).

I have one more point to address in [#105](https://github.com/pyomeca/pyomeca/issues/105).
I am waiting for @BKillen05 to approve [#107](https://github.com/pyomeca/pyomeca/issues/107).
