### 2020-08-20 4367e03
Format pyomeca as inline code
