### 2020-08-20 913de6c
Simplify sentence in statement of need
