### 2020-08-20 ccbb1b3
Add outline of the project and dev philosophy in the summary in response to #105
