### 2020-08-20 #2431 @romainmartinez
> Yes, I was OK with #105 The authors will have to update the manuscript based on our discussion.

Done in [PR #109](https://github.com/pyomeca/pyomeca/pull/109) @mitkof6.
