### 2020-08-20 pyomeca/pyomeca#107 @BKillen05
Thank you for the replies and the additional details provided, as well the links to the relevant documentation they were exactly the examples I was looking for.

From my point of view the changes and explanations you have given here sufficiently address my concerns
