### 2020-08-20 #2431 @romainmartinez
@Kevin-Mattheus-Moerman, with issue [#107](https://github.com/pyomeca/pyomeca/issues/107) just accepted and closed, [all issues](https://github.com/openjournals/joss-reviews/issues/2431#issuecomment-669236807) have been addressed 🎉 .

Perhaps @mitkof6 could confirm that the modifications made in the paper [here](https://github.com/pyomeca/pyomeca/pull/109) are suitable for him.
