### 2020-08-20 #2431 @Kevin-Mattheus-Moerman
@mitkof6 it looks like changes were made to the paper. Can you review them? :point_up: thanks!

@BKillen05 your only unticked box is "references". Can you comment on that or can that box be ticked now too?

@mitkof6, @BKillen05, if, after reviewe the changes made, you are happy to formulate a recommendation e.g. to accept this submission, please do so. Thanks!
