### 2020-08-21 #2431 @BKillen05
@Kevin-Mattheus-Moerman I have checked the references and now checked this box

As @mitkof6  , I am happy with the changes the authors have made to the documentation and also the manuscript.

Congratulations and good work from the authors
