### 2020-08-31 #2431 @Kevin-Mattheus-Moerman
@romainmartinez

I've reviewed your paper. Here are some final comments (fee free to disagree with these):

- [ ] Throughout this paper it seems to be the word "biomechanics" is used to refer to "musculoskeletal biomechanics", or at least biomechanics research featuring motion capture data. Given this, the following sentence (and some others) feels overly general: `While biomechanics branches into several subfields, the data used are remarkably similar.`. This may be true for musculoskeletal biomechanics research, however for image-based modelling and soft tissue biomechanics (a domain I work in) tables and scalar 1D data sets are rarer than multidimensional and multi-model image data (like 3D MRI, CT, and stereo-photography data). If you agree that the focus on musculoskeletal biomechanics is more accurate perhaps rephrase the first paragraph and perhaps wording throughout the text. This is only a recommendation, feel free to argue your way out of it.

- [ ] Please consider (suggestion, not required) to replace `not free of charge` by `proprietary`

- [ ] Capitalize `python` -> `Python` in the text

- [ ] Consider rephrasing `By offering a single, efficient and flexible implementation, pyomeca standardizes these procedures and avoids wasting valuable research time writing functions that have, sometimes, nothing to do with the research interests`, e.g. something like `By offering a single, efficient, and flexible implementation, pyomeca standardizes these procedures, freeing up valuable research time, thereby allowing researchers to focus on the scientific research questions at hand.`

- [ ] `...which is a step towards the adoption of programming in biomechanics`, Perhaps rephrase this sentence. This is perhaps too general again. Programming is used in biomechanics research quite intensively. It should be used more but this sentence seems confusing to me. Do you meant to say that these tools enhance reproducibility by enabling automated and code based analysis (as opposed to manual)?

- [ ] Should this `the xarray features set` be `the xarray feature set`?

- [ ] Spell out/introduce the EMG acronym

- [ ] Should this `skin markers position` be `skin marker positions`(check I think this occurs twice in the text)?

- [ ] In Figure 4 only the red, brown and purple lines can be distinguished. If this was intentional ignore this but if not consider changing the plotting order/type of visualization so more curves can be seen.

- [ ] You mention `numpy, scipy, matplotlib and xarray`, could you also cite these works?

If you implement changes please call `@whedon generate pdf` here to update the article proof.
