### 2020-08-31 0ac60bb
Rephrasing "By offering [...]"
