### 2020-08-31 04716e2
Move the EMG acronym up
