+++ docs/paper/paper.md
@@ -65,7 +65,7 @@ Each structure is associated with a specific biomechanical data type:
 - `Angles`: joint angles,
 - `Rototrans`: rototranslation matrix,
 - `Analogs`: generic signals such as electromyographic (EMG) and force signals or any other analog signals,
-- `Markers`: skin markers position.
+- `Markers`: skin markers positions.
 
 While there are technically dozens of functions implemented in `pyomeca`, one can generally group them into two distinct categories: object creation and data processing.
 
@@ -86,7 +86,7 @@ These methods can be categorized into filters (orange), signal processing (red),
 
 ## A Biomechanical Example: Electromyographic Pipeline
 
-`pyomeca` has documented examples for different biomechanical tasks such as getting Euler angles from a rototranslation matrix, creating a system of axes from skin markers position or setting a rotation or a translation.
+`pyomeca` has documented examples for different biomechanical tasks such as getting Euler angles from a rototranslation matrix, creating a system of axes from skin markers positions or setting a rotation or a translation.
 Another typical task concerns EMG data processing.
 Using `pyomeca`, one can easily extract (\autoref{fig:ex-1-raw}), process (\autoref{fig:ex-2-processed}) and visualize (\autoref{fig:ex-3-aggr}, \autoref{fig:ex-4-box} and \autoref{fig:ex-5-corr}) such data.
 
