### 2020-08-31 ddc1b3a
Cite librairies in acknowledgements
