+++ docs/paper/paper.md
@@ -46,7 +46,7 @@ Metadata are also needed to inform about important features of the experiment.\l
 As a Python library, `pyomeca` enables extraction, processing and visualization of biomechanical data for use in research and education.
 It is motivated by the need for simpler tools and more reproducible workflows allowing practitioners to focus on their specific interests and leaving `pyomeca` to handle the computational details for them.
 `pyomeca` builds on the core scientific Python packages, in particular `numpy` [@Walt2011-em], `scipy` [@Virtanen2020-zv], `matplotlib` [@Hunter2007-fv] and `xarray` [@Hoyer2017-sf].
-By providing labeled querying and computation, efficient algorithms and persistent metadata, the integration of `xarray` facilitates usability, which is a step towards the adoption of programming in biomechanics.
+By providing labeled querying and computation, efficient algorithms and persistent metadata, the integration of `xarray` facilitates usability, which seems to be relevant in a context where scientists do not always receive extensive training in programming.
 `xarray` is designed as a general-purpose library and tries to avoid including domain specific functionalities --- but inevitably, the need for more domain specific logic arises.
 `pyomeca` provides a biomechanics layer that supports specialized file formats (`c3d`, `mat`, `trc`, `sto`, `mot`, `csv` and `xlsx`) and implements signal processing and matrix manipulation routines commonly used in musculoskeletal biomechanics.
 We believe `pyomeca` is a solid foundation for more advanced research implementations and could, in the future, provide functions for processing and extracting meaningful clinical outputs from the raw data.
