### 2020-09-01 #2431 @romainmartinez
Thanks for your comments @Kevin-Mattheus-Moerman.

> - [x] Throughout this paper it seems to be the word "biomechanics" is used to refer to "musculoskeletal biomechanics", or at least biomechanics research featuring motion capture data. Given this, the following sentence (and some others) feels overly general: While biomechanics branches into several subfields, the data used are remarkably similar.. This may be true for musculoskeletal biomechanics research, however for image-based modelling and soft tissue biomechanics (a domain I work in) tables and scalar 1D data sets are rarer than multidimensional and multi-model image data (like 3D MRI, CT, and stereo-photography data). If you agree that the focus on musculoskeletal biomechanics is more accurate perhaps rephrase the first paragraph and perhaps wording throughout the text. This is only a recommendation, feel free to argue your way out of it.

You are absolutely right, done in [`45429fa`](https://github.com/romainmartinez/pyomeca/commit/45429fa3c1ed06146358e4aca2c5e611793497c5).

> - [x] Please consider (suggestion, not required) to replace not free of charge by proprietary

Done in [`b30b116`](https://github.com/romainmartinez/pyomeca/commit/b30b116c470994d142f66cee9f96045f2e822c41).

> - [x] Capitalize python -> Python in the text

Done in [`37b7a55`](https://github.com/romainmartinez/pyomeca/commit/37b7a551e7504589105d50f7f7efd9336ca64542).

> - [x] Consider rephrasing By offering a single, efficient and flexible implementation, pyomeca standardizes these procedures and avoids wasting valuable research time writing functions that have, sometimes, nothing to do with the research interests, e.g. something like By offering a single, efficient, and flexible implementation, pyomeca standardizes these procedures, freeing up valuable research time, thereby allowing researchers to focus on the scientific research questions at hand.

Done in [`0ac60bb`](https://github.com/romainmartinez/pyomeca/commit/0ac60bb490d4f64f82d32d9f5ecf89f9a903d655).

> - [x] ...which is a step towards the adoption of programming in biomechanics, Perhaps rephrase this sentence. This is perhaps too general again. Programming is used in biomechanics research quite intensively. It should be used more but this sentence seems confusing to me. Do you meant to say that these tools enhance reproducibility by enabling automated and code based analysis (as opposed to manual)?

Done in [`6660860`](https://github.com/romainmartinez/pyomeca/commit/66608605e1ea63e811b57abb4a58e53ebc698d62).
Is this sentence better: "By providing labeled querying and computation, efficient algorithms and persistent metadata, the integration of `xarray` facilitates usability, which seems to be relevant in a context where scientists do not always receive extensive training in programming."

> - [x]  Should this the xarray features set be the xarray feature set?

Done in [`331d680`](https://github.com/romainmartinez/pyomeca/commit/331d680e59c982c3dfd0416f9828b6658a7642c1).

> - [x] Spell out/introduce the EMG acronym

Done in [`04716e2`](https://github.com/romainmartinez/pyomeca/commit/04716e2dfddb9f43ef982abab31247b325dfe27e).

> - [x] Should this skin markers position be skin marker positions(check I think this occurs twice in the text)?

Done in [`51805d7`](https://github.com/romainmartinez/pyomeca/commit/51805d711e305aabc9049d3d2a368bb25d450167).

> - [x] In Figure 4 only the red, brown and purple lines can be distinguished. If this was intentional ignore this but if not consider changing the plotting order/type of visualization so more curves can be seen.

The curves intentionally overlap to show the raw data without taking up too much space. The purpose of this figure is to show that the data is raw, rather than to identify individual curves.
The following figure (5) shows the individual curves.

> - [x] You mention numpy, scipy, matplotlib and xarray, could you also cite these works?

Done in [`ddc1b3a`](https://github.com/romainmartinez/pyomeca/commit/ddc1b3a1546d2cd9a6a4ebb2c1d82ca4a4fa8b9a).
Thanks for the heads up, these libraries were cited in the rest of the paper except in the acknowledgements section.
