### 2020-09-02 #2431 @Kevin-Mattheus-Moerman
@romainmartinez I've checked my requested changes and all seems fine now. Below are final steps.

- [ ] Can you archive a copy of the software on ZENODO and report back here with the DOI of the archived version (do not alter or set the DOI yourself, please accept the ZENODO generated DOI)? Please ensure that the title and author list is the same as the JOSS paper for this ZENODO archived version (you may need to edit this meta-data manually).

- [ ] Can you inform me about the version tag for the published work. Is it still at `2020.0.1` or has it moved on?
