### 2020-09-02 #2431 @romainmartinez
Thanks @Kevin-Mattheus-Moerman.

> Can you archive a copy of the software on ZENODO and report back here with the DOI of the archived version (do not alter or set the DOI yourself, please accept the ZENODO generated DOI)? Please ensure that the title and author list is the same as the JOSS paper for this ZENODO archived version (you may need to edit this meta-data manually).

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4012118.svg)](https://doi.org/10.5281/zenodo.4012118)

> Can you inform me about the version tag for the published work. Is it still at 2020.0.1 or has it moved on?

Yes, the current version is still `2020.0.1`.

> Your paper is about to be accepted in JOSS. I recommend that you review the paper yourself once more, and that you also check the author list, affiliations, and the acknowledgements section for accuracy. (call @whedon generate pdf here to update the paper).

I reread the paper, checked the author list, affiliations, and the acknowledgements section for accuracy, and all seems fine.
