### 2020-09-02 #2431 @Kevin-Mattheus-Moerman
@romainmartinez Minor point, the paper title is `pyomeca: An Open-Source Framework for Biomechanical Analysis`, while the archive is: `Pyomeca: a Python toolbox for biomechanics analysis`.

- [ ] Can you update the archive title to match that of the paper (fix use of capital letters)
