### 2020-09-02 #2431 @romainmartinez
> Can you update the archive title to match that of the paper (fix use of capital letters)

@Kevin-Mattheus-Moerman, It is done. Sorry I was not sure if the DOI was attached to the GitHub repo or the paper.
