+++ docs/paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Hoyer2017-sf,
-title = "xarray: {N-D} labeled Arrays and Datasets in Python",
+title = "{xarray}: {N-D} labeled Arrays and Datasets in {Python}",
 author = "Hoyer, Stephan and Hamman, Joseph J",
 journal = "Journal of Open Research Software",
 volume =  5,
@@ -11,7 +11,7 @@ doi = "10.5334/jors.148"
 }
 
 @ARTICLE{Virtanen2020-zv,
-title = "{SciPy} 1.0: fundamental algorithms for scientific computing in Python",
+title = "{SciPy} 1.0: fundamental algorithms for scientific computing in {Python}",
 author = "Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and van der Walt, St{\'e}fan J and Brett, Matthew and Wilson, Joshua and Millman, K Jarrod and Mayorov, Nikolay and Nelson, Andrew R J and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C J and Polat, {\.I}lhan and Feng, Yu and Moore, Eric W and VanderPlas, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E A and Harris, Charles R and Archibald, Anne M and Ribeiro, Ant{\^o}nio H and Pedregosa, Fabian and van Mulbregt, Paul and {SciPy 1.0 Contributors}",
 journal = "Nat. Methods",
 volume =  17,
@@ -32,7 +32,7 @@ booktitle = "{P}roceedings of the 9th {P}ython in {S}cience {C}onference",
 author = "McKinney, Wes",
 publisher = "SciPy",
 pages = "56--61",
-series = "Proceedings of the Python in Science Conference",
+series = "{Proceedings of the Python in Science Conference}",
 year =  2010,
 conference = "Python in Science Conference",
 location = "Austin, Texas",
@@ -80,7 +80,7 @@ doi = "10.1016/j.cmpb.2016.11.007"
 }
 
 @ARTICLE{Muller2019-cx,
-title = "{CusToM}: a Matlab toolbox for musculoskeletal simulation",
+title = "{CusToM}: a {Matlab} toolbox for musculoskeletal simulation",
 author = "Muller, Antoine and Pontonnier, Charles and Puchaud, Pierre and Dumont, Georges",
 journal = "JOSS",
 volume =  4,
@@ -133,7 +133,7 @@ doi = "10.1016/j.simpat.2006.09.001"
 
 @misc{ezc3d,
 author = {Michaud, Benjamin and Begon, Mickael},
-title = {EZC3D: Easy to use C3D reader/writer in C++, Python and Matlab},
+title = {EZC3D: Easy to use {C3D} reader/writer in {C++, Python and Matlab}},
 year = {2020},
 publisher = {GitHub},
 journal = {GitHub repository},
