### 2020-09-03 #2431 @kthyng
Hi @romainmartinez, I am the EIC on rotation this week and will help finish publishing your paper. Your references were not capitalized all correctly. Please check the edits I made [here](https://github.com/pyomeca/pyomeca/pull/112) and if you agree, merge them in.
