### 2020-09-03 #2431 @romainmartinez
> Hi @romainmartinez, I am the EIC on rotation this week and will help finish publishing your paper. Your references were not capitalized all correctly. Please check the edits I made here and if you agree, merge them in.

Hi @kthyng and thanks for the edits. I just merged them.
